w3c-reports/12092014/mac-wk2/runned-tests/cors/status-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?1&origin=none&code=400. Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?5&headers=x-custom&preflight=400. Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?3&headers=x-custom&code=400. Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?2&origin=none&preflight=200. Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?4&headers=x-custom&code=400&preflight=200. Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
The returned status code in different scenarios

   allowed  preflight  response  | status |
   -------  ---------  --------  | ------ |
 1      no          x       400  |      0 |
 2      no        200         x  |      0 |
 3     yes          x       400  |    400 |
 4     yes        200       400  |    400 |
 5     yes        400         x  |      0 |

PASS 1. CORS disallowed, response status 400. 
PASS 2. CORS disallowed, preflight status 200. 
FAIL 3. CORS allowed, response status 400. assert_unreached: error event Reached unreachable code(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_unreached@http://localhost:8800/resources/testharness.js:692:16
http://localhost:8800/cors/status.htm:48:37
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
FAIL 4. CORS allowed, preflight status 200, response status 400. assert_unreached: error event Reached unreachable code(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_unreached@http://localhost:8800/resources/testharness.js:692:16
http://localhost:8800/cors/status.htm:48:37
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
PASS 5. CORS allowed, preflight status 400. 
